Checking for rook color when setting castling
Checking for rook color when setting castling
In Chess960 we can have legal positions with
opponent rook in A or H file and with castling
available, for instance:

4k3/pppppppp/8/8/8/8/PPPPPPPP/rR2K3 w Q - 0 1

In those cases we pick up the wrong rook when
setting castling.

Fix it by checking the color of the rook.

Bug reported by Matthew Lai.

No functional change.
master (#357)
commit 2795aedbc3710287448bad058c6077920066ad30 1 parent 8e95c39
@mcostalba mcostalba authored